Create Octopath Traveler Companion Guide Project
I compiled this project together as a companion guide to the stats of each of the playable characters in the Square Enix game Octopath Traveler.  Running this code will create a tibble in R tidyverse that contains the name of each character, the amount of experience needed for each level, and their various stats at each level. It was done using data from the Octopath Traveler fandom wiki as well as data gathered by Reddit user u/shelagog at July 19, 2019. It uses tools from tidyverse, rvest, and googlesheets4. It will require the user to log into their google drive in order to be usable. Other than that, one should be able to simply run the code in Rstudio, then access the tibble called All_Char_Stats to access this information. Thank you, and enjoy.