Create Alone Survival Game
This is the code for a basic survival game created using C++. The gameplay is simple; the game will provide a list of options and the user will type in the letter corresponding to the desired option (just make sure not to input the wrong option). The game ends once the user gets the "survived" ending. There is also an easter egg, for which a clue is given in the prelude. Enjoy.